requirements.txt
fastapi
pydantic 
paramiko
uvicorn
python-dotenv
pythonping